`evaluation` - renamed `excel_sheet_defines` to `excel_sheet_utils`
