+++ srpb_evaluation/scripts/create_latex_table_from_results.py
@@ -5,7 +5,7 @@ Dependency:
 sudo apt install python3-openpyxl
 '''
 
-import excel_sheet_defines
+import excel_sheet_utils
 import json
 import sys
 
@@ -22,10 +22,10 @@ def load_data_from_excel(path: Path) -> Dict[str, Dict[str, float]]:
     # NOTE: data_only allows to read values instead of formulas
     wb = load_workbook(filename = str(path), data_only=True)
     # access specific sheet
-    sheet = wb[excel_sheet_defines.SHEET_NAME]
+    sheet = wb[excel_sheet_utils.SHEET_NAME]
     # upper left cell - cell that initiates results table
-    col_init = excel_sheet_defines.RESULT_INIT_COL
-    row_init = excel_sheet_defines.RESULT_INIT_ROW
+    col_init = excel_sheet_utils.RESULT_INIT_COL
+    row_init = excel_sheet_utils.RESULT_INIT_ROW
 
     col_metric_ids = col_init
